prevent  event commands from being used in private messages

+++ cogs/events.py
@@ -1,6 +1,7 @@
 from discord.ext import commands
 from db.dbase import DBase
 from datetime import datetime
+from utils.messages import delete_all
 import discord
 import asyncio
 import re
@@ -23,21 +24,25 @@ class Events:
     @event.command(pass_context=True)
     async def create(self, ctx):
 
+        if ctx.message.channel.is_private:
+            return await self.bot.say(ctx.message.author.mention
+                                      + ": That command is not supported in a direct message.")
+
         await self.bot.say(ctx.message.author.mention + ": Enter event title")
         msg = await self.bot.wait_for_message(author=ctx.message.author)
         title = msg.content
 
-        await self.bot.say(ctx.message.author.mention
-                           + ": Enter event description (type 'none' for no description)")
+        await self.bot.say(ctx.message.author.mention + ": Enter event description (type 'none' for no description)")
         msg = await self.bot.wait_for_message(author=ctx.message.author)
         description = ''
         if msg.content.upper() != 'NONE':
             description = msg.content
 
         start_time = None
+        time_msg_count = 0
         while not start_time:
-            await self.bot.say(ctx.message.author.mention
-                               + ": Enter event time (YYYY-MM-DD HH:MM AM/PM)")
+            time_msg_count += 3
+            await self.bot.say(ctx.message.author.mention + ": Enter event time (YYYY-MM-DD HH:MM AM/PM)")
             msg = await self.bot.wait_for_message(author=ctx.message.author)
             start_time_str = msg.content
             start_time_format = '%Y-%m-%d %I:%M %p'
@@ -50,43 +55,45 @@ class Events:
         msg = await self.bot.wait_for_message(author=ctx.message.author)
         time_zone = msg.content.upper()
 
+        # Add event to the database
         with DBase() as db:
             db.create_event(title, start_time, time_zone, ctx.message.server.id, description)
         await self.bot.say(ctx.message.author.mention
                            + ": Event has been created! "
                            + "The list of upcoming events will be updated momentarily.")
+
+        # Clean up messages and update events channel
         await asyncio.sleep(4)
+        await self.bot.purge_from(ctx.message.channel, limit=(7 + time_msg_count), check=delete_all)
         await self.list_events(ctx.message.server)
 
 
     @event.command(pass_context=True)
     async def delete(self, ctx, event_id=None):
 
+        if ctx.message.channel.is_private:
+            return await self.bot.say(ctx.message.author.mention
+                                      + ": That command is not supported in a direct message.")
+
         # Attempt to delete the event with the given event_id
-        # If successful, update the events channel
-        msg = None
         deleted = False
         if event_id is not None:
             with DBase() as db:
                 affected_count = db.delete_event(event_id)
                 if affected_count > 0:
                     deleted = True
+            if deleted:
+                await self.bot.say(ctx.message.author.mention
+                                   + ": Event successfuly deleted. The list of upcoming "
+                                   + "events will be updated momentarily.")
+            else:
+                await self.bot.say(ctx.message.author.mention + ": That event doesn't exist.")
         else:
-            msg = await self.bot.say(ctx.message.author.mention
-                                     + ": An event ID must be specified! (Eg. '!event delete 117')")
-            await asyncio.sleep(4)
-            await self.bot.delete_message(msg)
-            await self.bot.delete_message(ctx.message)
-            return
-        if deleted:
-            msg = await self.bot.say(ctx.message.author.mention
-                                     + ": Event successfuly deleted. The list of upcoming "
-                                     + "events will be updated momentarily.")
-        else:
-            msg = await self.bot.say(ctx.message.author.mention + ": That event doesn't exist.")
-        await asyncio.sleep(4)
-        await self.bot.delete_message(msg)
-        await self.bot.delete_message(ctx.message)
+            await self.bot.say(ctx.message.author.mention
+                               + ": An event ID must be specified! (Eg. '!event delete 117')")
+
+        # Clean up messages and update the events channel if necessary
+        await self.bot.purge_from(ctx.message.channel, limit=2, check=delete_all)
         if deleted:
             await self.list_events(ctx.message.server)
 
@@ -96,14 +103,13 @@ class Events:
         # Clear the event channel and display all upcoming events
         events = None
         events_channel = await self.get_events_channel(server)
+        await self.bot.purge_from(events_channel, limit=999, check=delete_all)
         with DBase() as db:
             events = db.get_events(server.id)
         if len(events) != 0:
-            await self.bot.purge_from(events_channel, limit=999, check=self.check_delete)
             for row in events:
                 embed_msg = discord.Embed(color=discord.Colour(3381759))
-                embed_msg.set_footer(text="Use '!event delete "
-                                           + str(row[0]) + "' to remove this event")
+                embed_msg.set_footer(text="Use '!event delete " + str(row[0]) + "' to remove this event")
                 embed_msg.title = row[1]
                 if row[2]:
                     embed_msg.description = row[2]
@@ -117,8 +123,7 @@ class Events:
 
     async def on_reaction_add(self, reaction, user):
 
-        # If reaction is indicating event attendance,
-        # update the database and remove the reaction
+        # If reaction is indicating event attendance, update the database
         channel_name = reaction.message.channel.name
         author = reaction.message.author
         num_embeds = len(reaction.message.embeds)
@@ -175,8 +180,3 @@ class Events:
     # wait_for_reaction() to work properly
     async def on_server_available(self, server):
         await self.list_events(server)
-
-
-    # Helper function used to delete all messages
-    def check_delete(self, m):
-        return True;
